Don't crash when failing to create HTML tab
When for some reason "Swt::Browser.new" throws an exception, don't
leave a half-created tab item that will cause the application to crash
completely if the close button is clicked.

The implementation makes it very difficult to handle such an exceiption,
because the browser is created in an observer listener, making it
difficult to know what happened before the exception was thrown.

Ideally, the exception should be handled inside the listener function,
and no listener should ever throw. But in this case, there is code
outside the listener that depends on this listener adding the "browser"
attribute. So if we catch the exception inside the callback, a new
exception will occur when returning.